feat/#128: 컴포넌트 정리 API 연동 및 data fetch는 ExpTimeLine에서, UI는 Timeline에서 구현
