+++ app/(main)/dashboard/page.tsx
@@ -2,7 +2,7 @@ import ProfileCard from './components/ProfileCard';
 import ExpHistory from './components/ExpHistory';
 import Scrap from './components/Scrap';
 import ExpTimeLine from './components/ExpTimeLine';
-import { getExpHistory } from '@/apis/dashboard';
+import { getExpHistory, getExpTimeline } from '@/apis/dashboard';
 import Footer from '@/app/components/Footer';
 import {
   LazyJobRatioContainer,
@@ -18,6 +18,33 @@ export default async function DashboardPage() {
     new Date().getFullYear(),
     new Date().getMonth() + 1
   );
+
+  function getStartAndEndLines(): {
+    start: Date;
+    end: Date;
+    startLine: string;
+    endLine: string;
+  } {
+    const now = new Date();
+
+    // endLine: 이번 달의 마지막 날
+    const end = new Date(now.getFullYear(), now.getMonth() + 1, 1);
+
+    // startLine: 두 달 전의 첫째 날
+    const start = new Date(now.getFullYear(), now.getMonth() - 3, 2);
+
+    const toString = (date: Date) => date.toISOString().split('T')[0]; // yyyy-mm-dd 형식
+
+    return {
+      start: start,
+      end: end,
+      startLine: toString(start),
+      endLine: toString(end),
+    };
+  }
+
+  const { start, end, startLine, endLine } = getStartAndEndLines();
+  const expTimeline = await getExpTimeline(startLine, endLine);
 
   return (
     <div className="min-h-screen py-6 px-14">
@@ -39,7 +66,7 @@ export default async function DashboardPage() {
         </div>
         <div className="flex flex-grow flex-wrap lg:flex-nowrap gap-4 h-auto">
           <div className="flex-[2.5] bg-gray-800 rounded-[23px] pt-8 pb-4 px-8">
-            <ExpTimeLine />
+            <ExpTimeLine start={start} end={end} expTimeline={expTimeline} />
           </div>
           <div className="flex-[1] bg-gray-800 rounded-[23px] pt-8 pb-4 px-8">
             <ExpHistory expHistory={expHistory} />
